feat: provider-failure-kiro-parser-003 - Preserve Kiro invocation and resume behavior
